requirements.txt
allure-pytest==2.12.0
pytest==7.2.1
pydantic==1.10.4
python-dotenv==0.21.1
selenium==4.8.0
selenium-wire==5.1.0
webdriver-manager==3.8.5